% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robservatory-package.R
\docType{package}
\name{robservatory-package}
\alias{robservatory}
\alias{robservatory-package}
\title{robservatory: shared R Observatory utilities}
\description{
Internal utilities shared across the r-observatory pipelines. The initial
content resolves an R package name to its CRAN or Bioconductor identity.
}
\author{
\strong{Maintainer}: R Observatory \email{james.balamuta@gmail.com}

Authors:
\itemize{
  \item R Observatory \email{james.balamuta@gmail.com}
}

}
\keyword{internal}
